backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
ecdsa==0.19.1
fastapi==0.117.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
httplib2==0.31.0
httptools==0.6.4
idna==3.10
packaging==25.0
passlib==1.7.4
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pyparsing==3.2.5
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlmodel==0.0.25
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==1.1.0
websockets==15.0.1
